NEW segmentation UI

+++ core/src/main/java/im/dacer/kata/core/util/ViewUtil.kt
@@ -15,8 +15,8 @@ object ViewUtil {
         return Math.round(dp * density)
     }
 
-    fun sp2px(spValue: Float): Int {
+    fun sp2px(spValue: Float): Float {
         val fontScale = Resources.getSystem().displayMetrics.scaledDensity
-        return (spValue * fontScale + 0.5f).toInt()
+        return (spValue * fontScale + 0.5f)
     }
 }
